Implement `Reader.coroutine`
This function, similar to the coroutine function that you can find in
Bluebird, allow to pass a function generator and yield readers
intances.

Yielded values are transformed with Reader.of, and therefore, you can
also yield promises, and the generator function will wait until it is
resolved and continue.

Added a few tests cases to cover some of the complicated cases with
error handling.